Extract new tests to SqlTests.IgniteDbDataReader.cs
